pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "occ-mrd-runner"
version = "1.1.0"
description = "OCC (Operational Consistency Compiler) runtime + MRD suite runner"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Marco A. Isaac" }]
license = { text = "Apache-2.0" }
keywords = ["OCC", "ILSC", "MRD", "reproducibility"]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
  "jsonschema>=4.18",
  "reportlab>=3.6",
  "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5",
  "mypy>=1.7",
  "pre-commit>=3.5",
]

# Optional quality-of-life extras
cli = [
  "rich>=13",
]

docs = [
  "mkdocs-material>=9",
]

[project.urls]
Homepage = "https://github.com/MarcoAIsaac/OCC"
Repository = "https://github.com/MarcoAIsaac/OCC"
Issues = "https://github.com/MarcoAIsaac/OCC/issues"

[project.scripts]
occ = "occ.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["occ*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
  "ILSC_MRD_suite_15_modulos_CANON",
  "ILSC_MRD_suite_extensions",
  "docs",
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
